Move gas_subsidies test to its own folder.
